backend/requirements.txt
fastapi
uvicorn
tensorflow
keras
nltk
numpy
python-dotenv
google-generativeai
